Fixes Hyper-V agent port disconnect
On Windows / Hyper-V Server 2008 R2, when a switch port
have to be unbound because the VM using it was removed,
DisconnectSwitchPort will fail, returning an error code
and a HyperVException is raised.

If the VM's VNIC has been removed, disconnecting the
switch port is no longer necessary.

This issue is not present in Windows / Hyper-V Server 2012,
as the switch ports are cleaned automatically.

Change-Id: I729753c15c51c86655651b132fcf8eab78884eb5