Deprecate HyperVNeutronPlugin
HyperVNeutronPlugin is hereby deprecated, as it is outdated
and it is no longer needed, thanks to Ml2Plugin. All of its
users should switch to Ml2Plugin and use the migration
script, if necessary.

This commit adds the method to migrate their database
tables for HyperVNeutronPlugin to Ml2 in the script
neutron/db/migration/migrate_to_ml2.py.

Partially-implements: blueprint core-vendor-decomposition

Change-Id: If02c4761d9653dcd81fc0043ecce86a3b149dd99